requirements.txt
flask
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-flask==0.43b0
